+++ Support/lib/helpers.rb
@@ -14,8 +14,10 @@ end
 
 module Helpers
   include Constants
+
   extend Storage
-  
+  extend Logging  
+
   module_function
 
   def pluralize(n, singular, plural=nil)
@@ -51,12 +53,16 @@ module Helpers
   def set_markers(mark, errors_list)
     errors_list.each do |line_number, errors|
       messages = []
-
+      
       errors.each do |data|
-        messages << "#{data[:message]}"
+        if data.has_key?(:file)
+          if TM_FILEPATH =~ /#{data[:file]}$/
+            messages << "#{data[:message]}"
+          end
+        end
       end
 
-      set_marker(mark, line_number, messages.join("\n"))
+      set_marker(mark, line_number, messages.join("\n")) if messages.size > 0
     end
   end
 
@@ -119,39 +125,63 @@ module Helpers
     return sprintf("%0#{padding}d", line_number)
   end
   
-  def boxify_errors(errors, line_count)
+  def boxify_errors(errors, lines_count)
     messages = []
     go_to_errors = []
 
     errors_size = errors.values.map { |arr| arr.size }.inject(0) { |sum, count| sum + count }
-    
 
-    messages << "⚠️ Found #{errors_size} #{pluralize(errors_size, "error")}! ⚠️\n"
-    messages << "🔍 Use Option ( ⌥ ) + G to jump error line!"
-    
-    errors.each do |error_line, errs|
-      errs.sort_by{|err| err[:line_number]}.each do |err|
-        messages << "  - #{err[:line_number]} -> #{err[:message]}"
+    messages << "⚠️ #{errors_size} #{pluralize(errors_size, "error")} were found in total! ⚠️"
+    messages << TOOLTIP_BORDER_CHAR * TOOLTIP_LINE_LENGTH.to_i
+    messages << ''
+
+    sorted_errors = errors.sort_by { |key, _| key }
+    any_external_file_error = false
+
+    popup_messages = []
+    sorted_errors.each do |error_item|
+      line_number = error_item[0]
+      line_errors = error_item[1]
+      line_errors.sort_by{|err| err[:line_number]}.each do |err|
+        fmt_ln = pad_number(lines_count, err[:line_number])
+        fmt_cn = pad_number(lines_count, err[:column_number])
         
-        fmt_ln = pad_number(line_count, err[:line_number])
-        fmt_cn = pad_number(line_count, err[:column_number])
-        go_to_errors << "#{fmt_ln}:#{fmt_cn} | #{err[:message]}"
+        other_filename = ''
+        if err.has_key?(:file)
+          if TM_FILEPATH =~ /#{err[:file]}$/
+            go_to_errors << "#{fmt_ln}:#{fmt_cn} | #{err[:message]}"
+          else
+            any_external_file_error = true
+            other_filename = "👽 in: [#{err[:file]}]: "
+          end
+        end
+        
+        popup_messages << "  - #{other_filename}#{fmt_ln} -> #{err[:message]}"
       end
     end
-    
-    create_storage(go_to_errors, true) if go_to_errors
-    messages.join("\n")
+    if go_to_errors.size > 0
+      create_storage(go_to_errors, true)
+      messages << "🔍 Use Option ( ⌥ ) + G to jump error line!"
+    end
+
+    messages << "👽 Error found in different file!" if any_external_file_error
+    messages << ""
+
+    messages.concat(popup_messages).join("\n")
   end
   
   def organize_errors(errors)
     errs = {}
 
     errors.each do |error|
+      logger.info "original error: #{error.inspect}"
       case error
       when /^\((.*)\):(.*):(\d+):(\d+):\s?(.*)$/
         error_type, file, line_number, column_number, message = $1, $2, $3.to_i, $4.to_i, $5
+        logger.fatal "error_type: #{error_type} - file: #{file}"
         errs[line_number] = [] unless errs.has_key?(line_number)
         err = {
+          :file => file.sub(/^(vet|shadow): ?\.\//, ''),
           :line_number => line_number,
           :column_number => column_number,
           :type => error_type,
@@ -162,6 +192,7 @@ module Helpers
         error_type, file, line_number, column_number, message = $1, $2, $3.to_i, $4.to_i, $5
         errs[line_number] = [] unless errs.has_key?(line_number)
         err = {
+          :file => file.sub(/^\.\//, ''),
           :line_number => line_number,
           :column_number => column_number,
           :type => error_type,
@@ -178,20 +209,7 @@ module Helpers
         }
         errs[0] << err
       end
-      # if error =~ /^\((.*)\):(.*):(\d+):(\d+):\s?(.*)$/
-      #   error_type, file, line_number, column_number, message = $1, $2, $3.to_i, $4.to_i, $5
-      #   errs[line_number] = [] unless errs.has_key?(line_number)
-      #
-      #   err = {
-      #     :line_number => line_number,
-      #     :column_number => column_number,
-      #     :type => error_type,
-      #     :message => "[#{error_type}]: #{message}",
-      #   }
-      #   errs[line_number] << err
-      # end
     end
-    
     errs
   end