+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `fieldalignment` autofix
 - Add go tools installer with: option + I
 - Add go to error line with: option + G
+- Add `golangci-lint` integration
 
 **2025-02-28**